Implemented physics hacks

+++ src/core/util.c
@@ -203,8 +203,8 @@ float lerp(float a, float b, float t)
  */
 float lerp_angle(float alpha, float beta, float t)
 {
-    v2d_t a = v2d_new(cos(alpha), sin(alpha));
-    v2d_t b = v2d_new(cos(beta), sin(beta));
+    v2d_t a = v2d_new(cosf(alpha), sinf(alpha));
+    v2d_t b = v2d_new(cosf(beta), sinf(beta));
     v2d_t c = v2d_lerp(a, b, t);
     return atan2(c.y, c.x);
 }
